+++ package.json
@@ -36,7 +36,6 @@
     "query-string": "^5.1.0",
     "rc-slider": "^8.3.2",
     "react": "15.4.2",
-    "react-beautiful-dnd": "^6.0.2",
     "react-dom": "15.4.2",
     "react-modal": "^3.0.4",
     "react-router-dom": "^4.2.2",
